% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{activateWindow}
\alias{activateWindow}
\title{Activate window}
\usage{
activateWindow(x)
}
\arguments{
\item{x}{Name of the window to activate.}
}
\description{
Activates the desired window by name
}
\examples{
\dontrun{activateWindow("Notepad")}
}
