% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key2code.R
\name{key2code}
\alias{key2code}
\title{Map from keys to codes}
\usage{
key2code()
}
\value{
\code{data.frame} map from special keys to codes understood by \code{SendKeys}.
}
\description{
This convenience utility can be called to view the mapping from special
keys to codes accepted by the \code{sendKeys} library.
}
\examples{
key2code()
}
