% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sendKeys}
\alias{sendKeys}
\title{Send keys to active window}
\usage{
sendKeys(x)
}
\arguments{
\item{x}{Single string containing the key sequence to send, possibly
including modifiers. See the examples.}
}
\description{
Sends keys to the active window. Beware that other processes
may affect what the active window might be while the key-strokes are sent.
Furthermore, the active window can be switched during a single call to
\code{sendKeys}.
}
\examples{
\dontrun{sendKeys("{DELAY=50}@rnotepad{ENTER}")}
\dontrun{sendKeys("{DELAY=50}@rchrome{ENTER}{DELAY=300}m{DELAY=50}irai-solutions^{ENTER}")}
}
\references{
\url{https://www.codeproject.com/Articles/6819/SendKeys-in-C}
}
