% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsendkeys-package.R
\docType{package}
\name{winsendkeys-package}
\alias{winsendkeys}
\alias{winsendkeys-package}
\title{Windows Sendkeys Package for R}
\description{
Wraps C++ sendkeys library allowing to send keystroke sequences, execute shortcuts and switch between active windows.
Provides a couple of more low-level SendInput examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.mirai-solutions.com}
  \item \url{https://github.com/miraisolutions/winsendkeys}
  \item Report bugs at \url{https://github.com/miraisolutions/winsendkeys/issues}
}

}
\author{
\strong{Maintainer}: Roland Schmid \email{roland.schmid@mirai-solutions.com}

Other contributors:
\itemize{
  \item Mirai Solutions GmbH \email{info@mirai-solutions.com} [copyright holder]
  \item Elias Bachaalany (Sendkeys C++ library) [contributor, copyright holder]
}

}
