% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sendKeys}
\alias{sendKeys}
\title{Send keys to active window}
\usage{
sendKeys(x)
}
\examples{
sendKeys("{DELAY=50}@rnotepad{ENTER}")
sendKeys("{DELAY=50}@rchrome{ENTER}{DELAY=200}h{DELAY=50}ttp://mirai-solutions.ch{ENTER}")
sendKeys("{DELAY=50}@rchrome{ENTER}{DELAY=300}m{DELAY=50}irai-solutions^{ENTER}")
}
\references{
url{https://www.codeproject.com/Articles/6819/SendKeys-in-C}
}
